move threshold panel code out of AppExplore
